requirements.txt
GitPython
PyYaml